email-management/agent/requirements.txt
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
supabase>=2.0.0
google-generativeai>=0.3.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
